requirements.txt
annotated-types==0.6.0
anyio==4.2.0
audioread==3.0.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.0
fastapi==0.109.0
flask==3.0.1
fonttools==4.47.2
google-api-core==2.15.0
google-auth==2.27.0
google-cloud-core==2.4.1
google-cloud-storage==1.43.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lazy-loader==0.3
librosa==0.10.1
llvmlite==0.41.1
MarkupSafe==2.1.4
matplotlib==3.8.2
msgpack==1.0.7
numba==0.58.1
numpy==1.26.3
packaging==23.2
pillow==10.2.0
pip-autoremove==0.10.0
platformdirs==4.1.0
pooch==1.8.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic-core==2.14.6
pydub==0.25.1
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
starlette==0.35.1
tenacity==8.2.3
threadpoolctl==3.2.0
typing-extensions==4.9.0
urllib3==2.1.0
uvicorn==0.27.0
werkzeug==3.0.1
zipp==3.17.0
